+++ CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-12-08
+
 ### Added
 
 - Optional DNS resolver. When enabled this will bind UDP port 53 on the system.
   For now, this uses the system configured resolvers to resolve queries. In the
   future, this will be expanded to redirect queries for certain TLD's to alternative
   backend.
+
+### Changed
+
+- The Quic connection type now uses quic datagrams to transport __data__ (packets
+  coming from the TUN device) to the peer. Protocol traffic is still sent over a
+  bidirectional Quic stream (which supports retransmits).
 
 ### Fixed
 
@@ -651,4 +659,5 @@ This is a breaking change, check the main README file for update info.
 [0.6.0]: https://github.com/threefoldtech/mycelium/compare/v0.5.7...v0.6.0
 [0.6.1]: https://github.com/threefoldtech/mycelium/compare/v0.6.0...v0.6.1
 [0.6.2]: https://github.com/threefoldtech/mycelium/compare/v0.6.1...v0.6.2
-[unreleased]: https://github.com/threefoldtech/mycelium/compare/v0.6.2...HEAD
+[0.7.0]: https://github.com/threefoldtech/mycelium/compare/v0.6.2...v0.7.0
+[unreleased]: https://github.com/threefoldtech/mycelium/compare/v0.7.0...HEAD